Added variable delay lstms.
