+++ egs/ami/s5/RESULTS_sdm
@@ -81,6 +81,12 @@ for x in exp/$mic/nnet2_online/*/decode*; do [ -d $x ] && [[ $x =~ "$1" ]] && gr
 %WER 44.2 | 14069 94507 | 61.3 25.8 12.9 5.5 44.2 70.7 | -24.180 | exp/sdm1/nnet3/lstm_sp_ld5/decode_dev/ascore_10/dev_hires_o4.ctm.filt.sys
 %WER 47.6 | 14034 89978 | 56.8 26.9 16.4 4.3 47.6 67.7 | -23.786 | exp/sdm1/nnet3/lstm_sp_ld5/decode_eval/ascore_10/eval_hires_o4.ctm.filt.sys
 
+#Variable-delay LSTM (Default LSTM recipe)
+#Total training time is 18:43:35
+%WER 43.8 | 14302 94500 | 61.7 25.3 13.0 5.5 43.8 69.4 | -24.338 | exp/sdm1/nnet3/lstm_sp_s4_lsd123_ld5/decode_dev/ascore_10/dev_hires_o4.ctm.filt.sys
+%WER 47.3 | 13268 89973 | 56.9 26.5 16.6 4.2 47.3 71.5 | -23.883 | exp/sdm1/nnet3/lstm_sp_s4_lsd123_ld5/decode_eval/ascore_10/eval_hires_o4.ctm.filt.sys
+
+
 #--------------------------------------------------------------------------------------------------------------------------------------------
 # Karel, JSALT 2015, (28.7.2015)
 # nnet1, MFCC-LDA-MLLT-DNN system (local/nnet/run_dnn_lda_mllt.sh),
