+++ egs/wsj/s5/steps/nnet3/lstm/make_configs.py
@@ -6,216 +6,9 @@ import argparse
 import sys
 import warnings
 import copy
+import imp
 
-# adds the input nodes and returns the descriptor
-def AddInputLayer(config_lines, feat_dim, splice_indexes=[0], ivector_dim=0):
-    components = config_lines['components']
-    component_nodes = config_lines['component-nodes']
-    output_dim = 0
-    components.append('input-node name=input dim=' + str(feat_dim))
-    list = [('Offset(input, {0})'.format(n) if n != 0 else 'input') for n in splice_indexes]
-    output_dim += len(splice_indexes) * feat_dim
-    if args.ivector_dim > 0:
-        components.append('input-node name=ivector dim=' + str(ivector_dim))
-        list.append('ReplaceIndex(ivector, t, 0)')
-        output_dim += ivector_dim
-    splice_descriptor = "Append({0})".format(", ".join(list))
-    print(splice_descriptor)
-    return {'descriptor': splice_descriptor,
-            'dimension': output_dim}
-
-def AddLdaLayer(config_lines, name, input, lda_file):
-    components = config_lines['components']
-    component_nodes = config_lines['component-nodes']
-
-    components.append('component name={0}_lda type=FixedAffineComponent matrix={1}'.format(name, lda_file))
-    component_nodes.append('component-node name={0}_lda component={0}_lda input={1}'.format(name, input['descriptor']))
-
-    return {'descriptor':  '{0}_lda'.format(name),
-            'dimension': input['dimension']}
-
-def AddAffineLayer(config_lines, name, input, output_dim, ng_affine_options = ""):
-    components = config_lines['components']
-    component_nodes = config_lines['component-nodes']
-
-    components.append("component name={0}_affine type=NaturalGradientAffineComponent input-dim={1} output-dim={2} {3}".format(name, input['dimension'], output_dim, ng_affine_options))
-    component_nodes.append("component-node name={0}_affine component={0}_affine input={1}".format(name, input['descriptor']))
-
-    return {'descriptor':  '{0}_affine'.format(name),
-            'dimension': output_dim}
-
-def AddAffRelNormLayer(config_lines, name, input, output_dim, ng_affine_options = ""):
-    components = config_lines['components']
-    component_nodes = config_lines['component-nodes']
-
-    components.append("component name={0}_affine type=NaturalGradientAffineComponent input-dim={1} output-dim={2} {3}".format(name, input['dimension'], output_dim, ng_affine_options))
-    components.append("component name={0}_relu type=RectifiedLinearComponent dim={1}".format(name, output_dim))
-    components.append("component name={0}_renorm type=NormalizeComponent dim={1}".format(name, output_dim))
-
-    component_nodes.append("component-node name={0}_affine component={0}_affine input={1}".format(name, input['descriptor']))
-    component_nodes.append("component-node name={0}_relu component={0}_relu input={0}_affine".format(name))
-    component_nodes.append("component-node name={0}_renorm component={0}_renorm input={0}_relu".format(name))
-
-    return {'descriptor':  '{0}_renorm'.format(name),
-            'dimension': output_dim}
-
-
-
-def AddSoftmaxLayer(config_lines, name, input):
-    components = config_lines['components']
-    component_nodes = config_lines['component-nodes']
-
-    components.append("component name={0}_log_softmax type=LogSoftmaxComponent dim={1}".format(name, input['dimension']))
-    component_nodes.append("component-node name={0}_log_softmax component={0}_log_softmax input={1}".format(name, input['descriptor']))
-
-    return {'descriptor':  '{0}_log_softmax'.format(name),
-            'dimension': input['dimension']}
-
-def AddOutputNode(config_lines, input):
-    components = config_lines['components']
-    component_nodes = config_lines['component-nodes']
-    component_nodes.append('output-node name=output input={0}'.format(input['descriptor']))
-
-
-def AddOutputNodeWithDelay(config_lines, input, label_delay):
-    components = config_lines['components']
-    component_nodes = config_lines['component-nodes']
-    component_nodes.append('output-node name=output input=Offset({0},{1})'.format(input['descriptor'], label_delay))
-
-def AddFinalLayer(config_lines, input, output_dim, ng_affine_options = ""):
-    prev_layer_output = AddAffineLayer(config_lines, "Final", input, output_dim, ng_affine_options)
-    prev_layer_output = AddSoftmaxLayer(config_lines, "Final", prev_layer_output)
-    AddOutputNodeWithDelay(config_lines, prev_layer_output, args.label_delay)
-
-
-def AddLstmLayer(config_lines,
-                 name, input, cell_dim,
-                 recurrent_projection_dim = 0,
-                 non_recurrent_projection_dim = 0,
-                 clipping_threshold = 1.0,
-                 norm_based_clipping = "false",
-                 ng_per_element_scale_options = "",
-                 ng_affine_options = ""):
-    assert(recurrent_projection_dim >= 0 and non_recurrent_projection_dim >= 0)
-    components = config_lines['components']
-    component_nodes = config_lines['component-nodes']
-
-    input_descriptor = input['descriptor']
-    input_dim = input['dimension']
-    name = name.strip()
-
-    if (recurrent_projection_dim == 0):
-        add_recurrent_projection = False
-        recurrent_projection_dim = cell_dim
-        recurrent_connection = "m_t"
-    else:
-        add_recurrent_projection = True
-        recurrent_connection = "r_t"
-    if (non_recurrent_projection_dim == 0):
-        add_non_recurrent_projection = False
-    else:
-        add_non_recurrent_projection = True
-
-    # Natural gradient per element scale parameters
-    ng_per_element_scale_options += " param-mean=0.0 param-stddev=1.0 "
-    # Parameter Definitions W*(* replaced by - to have valid names)
-    components.append("# Input gate control : W_i* matrices")
-    components.append("component name={0}_W_i-xr type=NaturalGradientAffineComponent input-dim={1} output-dim={2} {3}".format(name, input_dim + recurrent_projection_dim, cell_dim, ng_affine_options))
-    components.append("# note : the cell outputs pass through a diagonal matrix")
-    components.append("component name={0}_w_ic type=NaturalGradientPerElementScaleComponent  dim={1} {2}".format(name, cell_dim, ng_per_element_scale_options))
-
-    components.append("# Forget gate control : W_f* matrices")
-    components.append("component name={0}_W_f-xr type=NaturalGradientAffineComponent input-dim={1} output-dim={2} {3}".format(name, input_dim + recurrent_projection_dim, cell_dim, ng_affine_options))
-    components.append("# note : the cell outputs pass through a diagonal matrix")
-    components.append("component name={0}_w_fc type=NaturalGradientPerElementScaleComponent  dim={1} {2}".format(name, cell_dim, ng_per_element_scale_options))
-
-    components.append("#  Output gate control : W_o* matrices")
-    components.append("component name={0}_W_o-xr type=NaturalGradientAffineComponent input-dim={1} output-dim={2} {3}".format(name, input_dim + recurrent_projection_dim, cell_dim, ng_affine_options))
-    components.append("# note : the cell outputs pass through a diagonal matrix")
-    components.append("component name={0}_w_oc type=NaturalGradientPerElementScaleComponent  dim={1} {2}".format(name, cell_dim, ng_per_element_scale_options))
-
-    components.append("# Cell input matrices : W_c* matrices")
-    components.append("component name={0}_W_c-xr type=NaturalGradientAffineComponent input-dim={1} output-dim={2} {3}".format(name, input_dim + recurrent_projection_dim, cell_dim, ng_affine_options))
-
-
-    components.append("# Defining the non-linearities")
-    components.append("component name={0}_i type=SigmoidComponent dim={1}".format(name, cell_dim))
-    components.append("component name={0}_f type=SigmoidComponent dim={1}".format(name, cell_dim))
-    components.append("component name={0}_o type=SigmoidComponent dim={1}".format(name, cell_dim))
-    components.append("component name={0}_g type=TanhComponent dim={1}".format(name, cell_dim))
-    components.append("component name={0}_h type=TanhComponent dim={1}".format(name, cell_dim))
-
-    components.append("# Defining the cell computations")
-    components.append("component name={0}_c1 type=ElementwiseProductComponent input-dim={1} output-dim={2}".format(name, 2 * cell_dim, cell_dim))
-    components.append("component name={0}_c2 type=ElementwiseProductComponent input-dim={1} output-dim={2}".format(name, 2 * cell_dim, cell_dim))
-    components.append("component name={0}_m type=ElementwiseProductComponent input-dim={1} output-dim={2}".format(name, 2 * cell_dim, cell_dim))
-    components.append("component name={0}_c type=ClipGradientComponent dim={1} clipping-threshold={2} norm-based-clipping={3} ".format(name, cell_dim, clipping_threshold, norm_based_clipping))
-
-    # c1_t and c2_t defined below
-    component_nodes.append("component-node name={0}_c_t component={0}_c input=Sum({0}_c1_t, {0}_c2_t)".format(name))
-    c_tminus1_descriptor = "IfDefined(Offset({0}_c_t, -1))".format(name)
-
-    component_nodes.append("# i_t")
-    component_nodes.append("component-node name={0}_i1 component={0}_W_i-xr input=Append({1}, IfDefined(Offset({0}_{2}, -1)))".format(name, input_descriptor, recurrent_connection))
-    component_nodes.append("component-node name={0}_i2 component={0}_w_ic  input={1}".format(name, c_tminus1_descriptor))
-    component_nodes.append("component-node name={0}_i_t component={0}_i input=Sum({0}_i1, {0}_i2)".format(name))
-
-    component_nodes.append("# f_t")
-    component_nodes.append("component-node name={0}_f1 component={0}_W_f-xr input=Append({1}, IfDefined(Offset({0}_{2}, -1)))".format(name, input_descriptor, recurrent_connection))
-    component_nodes.append("component-node name={0}_f2 component={0}_w_fc  input={1}".format(name, c_tminus1_descriptor))
-    component_nodes.append("component-node name={0}_f_t component={0}_f input=Sum({0}_f1,{0}_f2)".format(name))
-
-    component_nodes.append("# o_t")
-    component_nodes.append("component-node name={0}_o1 component={0}_W_o-xr input=Append({1}, IfDefined(Offset({0}_{2}, -1)))".format(name, input_descriptor, recurrent_connection))
-    component_nodes.append("component-node name={0}_o2 component={0}_w_oc input={0}_c_t".format(name))
-    component_nodes.append("component-node name={0}_o_t component={0}_o input=Sum({0}_o1, {0}_o2)".format(name))
-
-    component_nodes.append("# h_t")
-    component_nodes.append("component-node name={0}_h_t component={0}_h input={0}_c_t".format(name))
-
-    component_nodes.append("# g_t")
-    component_nodes.append("component-node name={0}_g1 component={0}_W_c-xr input=Append({1}, IfDefined(Offset({0}_{2}, -1)))".format(name, input_descriptor, recurrent_connection))
-    component_nodes.append("component-node name={0}_g_t component={0}_g input={0}_g1".format(name))
-
-    component_nodes.append("# parts of c_t")
-    component_nodes.append("component-node name={0}_c1_t component={0}_c1  input=Append({0}_f_t, {1})".format(name, c_tminus1_descriptor))
-    component_nodes.append("component-node name={0}_c2_t component={0}_c2 input=Append({0}_i_t, {0}_g_t)".format(name))
-
-    component_nodes.append("# m_t")
-    component_nodes.append("component-node name={0}_m_t component={0}_m input=Append({0}_o_t, {0}_h_t)".format(name))
-
-    # add the recurrent connections
-    if (add_recurrent_projection and add_non_recurrent_projection):
-        components.append("# projection matrices : Wrm and Wpm")
-        components.append("component name={0}_W-m type=NaturalGradientAffineComponent input-dim={1} output-dim={2} {3}".format(name, cell_dim, recurrent_projection_dim + non_recurrent_projection_dim, ng_affine_options))
-        components.append("component name={0}_r type=ClipGradientComponent dim={1} clipping-threshold={2} norm-based-clipping={3} ".format(name, recurrent_projection_dim, clipping_threshold, norm_based_clipping))
-        component_nodes.append("# r_t and p_t")
-        component_nodes.append("component-node name={0}_rp_t component={0}_W-m input={0}_m_t".format(name))
-        component_nodes.append("dim-range-node name={0}_r_t_preclip input-node={0}_rp_t dim-offset=0 dim={1}".format(name, recurrent_projection_dim))
-        component_nodes.append("component-node name={0}_r_t component={0}_r input={0}_r_t_preclip".format(name))
-        output_descriptor = '{0}_rp_t'.format(name)
-        output_dim = recurrent_projection_dim + non_recurrent_projection_dim
-
-    elif add_recurrent_projection:
-        components.append("# projection matrices : Wrm")
-        components.append("component name={0}_Wrm type=NaturalGradientAffineComponent input-dim={1} output-dim={2} {3}".format(name, cell_dim, recurrent_projection_dim, ng_affine_options))
-        components.append("component name={0}_r type=ClipGradientComponent dim={1} clipping-threshold={2} norm-based-clipping={3} ".format(name, recurrent_projection_dim, clipping_threshold, norm_based_clipping))
-        component_nodes.append("# r_t")
-        component_nodes.append("component-node name={0}_r_t_preclip component={0}_Wrm input={0}_m_t".format(name))
-        component_nodes.append("component-node name={0}_r_t component={0}_r input={0}_r_t_preclip".format(name))
-        output_descriptor = '{0}_r_t'.format(name)
-        output_dim = recurrent_projection_dim
-
-    else:
-        components.append("component name={0}_r type=ClipGradientComponent dim={1} clipping-threshold={2} norm-based-clipping={3} ".format(name, cell_dim, clipping_threshold, norm_based_clipping))
-        component_nodes.append("component-node name={0}_r_t component={0}_r input={0}_m_t".format(name))
-        output_descriptor = '{0}_r_t'.format(name)
-        output_dim = cell_dim
-
-    return {
-            'descriptor': output_descriptor,
-            'dimension':output_dim
-            }
+nodes = imp.load_source('', 'steps/nnet3/components.py')
 
 def PrintConfig(file_name, config_lines):
     f = open(file_name, 'w')
@@ -224,14 +17,18 @@ def PrintConfig(file_name, config_lines):
     f.write("\n".join(config_lines['component-nodes']))
     f.close()
 
-def ParseSpliceString(splice_indexes, label_delay):
+def ParseSpliceString(splice_indexes, label_delay=None):
     ## Work out splice_array e.g. splice_array = [ [ -3,-2,...3 ], [0], [-2,2], .. [ -8,8 ] ]
     split1 = splice_indexes.split(" ");  # we already checked the string is nonempty.
     if len(split1) < 1:
         splice_indexes = "0"
 
-    left_context = -label_delay
-    right_context = label_delay
+    left_context=0
+    right_context=0
+    if label_delay is not None:
+        left_context = -label_delay
+        right_context = label_delay
+
     splice_array = []
     try:
         for i in range(len(split1)):
@@ -286,6 +83,9 @@ if __name__ == "__main__":
                         help="dimension of non-recurrent projection")
     parser.add_argument("--hidden-dim", type=int,
                         help="dimension of fully-connected layers")
+    # CLSTM options
+    parser.add_argument("--rates", type=str, default=None,
+                        help="the rates of operation in each clockwork-lstm")
 
     # Natural gradient options
     parser.add_argument("--ng-per-element-scale-options", type=str,
@@ -305,8 +105,12 @@ if __name__ == "__main__":
                         help="Directory to write config files and variables")
 
     # Delay options
-    parser.add_argument("--label-delay", type=int,
+    parser.add_argument("--label-delay", type=int, default=None,
                         help="option to delay the labels to make the lstm robust")
+
+    parser.add_argument("--lstm-delay", type=str, default=None,
+                        help="option to have different delays in recurrence for each lstm")
+
 
 
     print(' '.join(sys.argv))
@@ -327,10 +131,30 @@ if __name__ == "__main__":
         sys.exit("--num-lstm-layers has to be a positive integer")
     if (args.clipping_threshold < 0):
         sys.exit("--clipping-threshold has to be a non-negative")
+    if args.lstm_delay is None:
+        lstm_delay = [-1] * args.num_lstm_layers
+    else:
+        try:
+            lstm_delay = map(lambda x: int(x), args.lstm_delay.split())
+        except ValueError:
+            sys.exit("--lstm-delay has incorrect format value. Provided value is '{0}'".format(args.lstm_delay))
+        if len(lstm_delay) != args.num_lstm_layers:
+            sys.exit("--lstm-delay: Number of delays provided has to match --num-lstm-layers")
+    if args.rates is None:
+        rates = [1]
+    else:
+        try:
+            rates = map(lambda x: float(x), args.lstm_delay.split())
+            rates.sort()
+            # assert that lower rates are divisors of higher rates
+            for rate_index in len(rates):
+                for higher_rate in rates[rate_index+1:]:
+                    if not (higher_rate % rates[rate_index] == 0):
+                        sys.exit("--rates : higher rates should be divisible by all lower rates")
+        except ValueError:
+            sys.exit("--rates has incorrect format value. Provided value is '{0}'".format(args.rates))
 
-
-
-    parsed_splice_output = ParseSpliceString(args.splice_indexes.strip(),args.label_delay)
+    parsed_splice_output = ParseSpliceString(args.splice_indexes.strip(), args.label_delay)
     left_context = parsed_splice_output['left_context']
     right_context = parsed_splice_output['right_context']
     num_hidden_layers = parsed_splice_output['num_hidden_layers']
@@ -350,35 +174,36 @@ if __name__ == "__main__":
     config_lines = {'components':[], 'component-nodes':[]}
 
     config_files={}
-    prev_layer_output = AddInputLayer(config_lines, args.feat_dim, splice_indexes[0], args.ivector_dim)
+    prev_layer_output = nodes.AddInputLayer(config_lines, args.feat_dim, splice_indexes[0], args.ivector_dim)
 
     # Add the init config lines for estimating the preconditioning matrices
     init_config_lines = copy.deepcopy(config_lines)
     init_config_lines['components'].insert(0, '# Config file for initializing neural network prior to')
     init_config_lines['components'].insert(0, '# preconditioning matrix computation')
-    AddOutputNode(init_config_lines, prev_layer_output)
+    nodes.AddOutputNode(init_config_lines, prev_layer_output)
     config_files[args.config_dir + '/init.config'] = init_config_lines
 
-    prev_layer_output = AddLdaLayer(config_lines, "L0", prev_layer_output, args.config_dir + '/lda.mat')
+    prev_layer_output = nodes.AddLdaLayer(config_lines, "L0", prev_layer_output, args.config_dir + '/lda.mat')
 
     for i in range(args.num_lstm_layers):
-        prev_layer_output = AddLstmLayer(config_lines, "Lstm{0}".format(i+1), prev_layer_output, args.cell_dim,
+        prev_layer_output = nodes.AddLstmLayer(config_lines, "Lstm{0}".format(i+1), prev_layer_output, args.cell_dim,
                                          args.recurrent_projection_dim, args.non_recurrent_projection_dim,
                                          args.clipping_threshold, args.norm_based_clipping,
-                                         args.ng_per_element_scale_options, args.ng_affine_options)
+                                         args.ng_per_element_scale_options, args.ng_affine_options,
+                                         lstm_delay = lstm_delay[i])
         # make the intermediate config file for layerwise discriminative
         # training
-        AddFinalLayer(config_lines, prev_layer_output, args.num_targets, args.ng_affine_options)
+        nodes.AddFinalLayer(config_lines, prev_layer_output, args.num_targets, args.ng_affine_options, args.label_delay)
         config_files['{0}/layer{1}.config'.format(args.config_dir, i+1)] = config_lines
         config_lines = {'components':[], 'component-nodes':[]}
 
     for i in range(args.num_lstm_layers, num_hidden_layers):
-        prev_layer_output = AddAffRelNormLayer(config_lines, "L{0}".format(i+1),
+        prev_layer_output = nodes.AddAffRelNormLayer(config_lines, "L{0}".format(i+1),
                                                prev_layer_output, args.hidden_dim,
                                                args.ng_affine_options)
         # make the intermediate config file for layerwise discriminative
         # training
-        AddFinalLayer(config_lines, prev_layer_output, args.num_targets, args.ng_affine_options)
+        nodes.AddFinalLayer(config_lines, prev_layer_output, args.num_targets, args.ng_affine_options, args.label_delay)
         config_files['{0}/layer{1}.config'.format(args.config_dir, i+1)] = config_lines
         config_lines = {'components':[], 'component-nodes':[]}
